Added stats bar and made images get saved.

+++ PoeTradeDesktop/Schemes/Searching/_SearchResultItem/_Item/Extended.cs
@@ -7,12 +7,12 @@ namespace PoeTradeDesktop.Schemes.Searching._SearchResultItem._Item
         public Mods Mods { get; set; }
         public Hashes Hashes { get; set; }
 
-        public int Ar { get; set; }
-        public int Ev { get; set; }
-        public int Es { get; set; }
-        public int Dps { get; set; }
-        public int Pdps { get; set; }
-        public int Edps { get; set; }
+        public float Ar { get; set; }
+        public float Ev { get; set; }
+        public float Es { get; set; }
+        public float Dps { get; set; }
+        public float Pdps { get; set; }
+        public float Edps { get; set; }
 
         public bool Ar_aug { get; set; }
         public bool Ev_aug { get; set; }
